Refactor logger imports to use the utils directory for consistency across DNS provider modules
